pyborg/requirements.txt
aiohttp==3.6.2
appdirs==1.4.4
arrow==0.15.8
async-timeout==3.0.1
attrs==19.3.0
blurhash==1.1.4
bottle==0.12.18
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
decorator==4.4.2
defusedxml==0.6.0
discord.py==1.4.1
distlib==0.3.1
filelock==3.0.12
future==0.18.2
humanize==2.6.0
idna==2.10
importlib-metadata==1.7.0
irc==18.0.0
jaraco.classes==3.1.0
jaraco.collections==3.0.0
jaraco.functools==3.0.1
jaraco.logging==3.0.0
jaraco.stream==3.0.0
jaraco.text==3.2.0
Mastodon.py==1.5.1
more-itertools==8.4.0
multidict==4.7.6
oauthlib==3.1.0
packaging==20.4
pluggy==0.13.1
praw==6.5.1
prawcore==1.5.0
prompt-toolkit==3.0.6
py==1.9.0
pyborg @ file:///home/benjamin/PythonProjects/pyborg-1up
pyborg-custom==1.0.1
pyparsing==2.4.7
PySocks==1.7.1
python-dateutil==2.8.1
python-magic==0.4.18
PyTumblr==0.1.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
statsd==3.3.0
tabulate==0.8.7
tempora==4.0.0
toml==0.10.1
tox==3.19.0
tweepy==3.9.0
typing-extensions==3.7.4.2
update-checker==0.18.0
urllib3==1.25.10
venusian==3.0.0
virtualenv==20.0.31
wcwidth==0.2.5
websocket-client==0.57.0
yarl==1.5.1
zipp==3.1.0
